port DBConnectionOptions to use DI as well
