+++ packages/cli/src/commands/db/revert.ts
@@ -6,7 +6,7 @@ import { MigrationExecutor, DataSource as Connection } from '@n8n/typeorm';
 import { Command, Flags } from '@oclif/core';
 import { Logger } from 'n8n-core';
 
-import { getConnectionOptions } from '@/databases/config';
+import { DBConnectionOptions } from '@/databases/db-connection-options';
 import type { Migration } from '@/databases/types';
 import { wrapMigration } from '@/databases/utils/migration-helpers';
 
@@ -80,7 +80,7 @@ export class DbRevertMigrationCommand extends Command {
 
 	async run() {
 		const connectionOptions: ConnectionOptions = {
-			...getConnectionOptions(),
+			...Container.get(DBConnectionOptions).getOptions(),
 			subscribers: [],
 			synchronize: false,
 			migrationsRun: false,
