+++ packages/cli/src/databases/db-connection-options.ts
@@ -1,0 +1,146 @@
+import { GlobalConfig } from '@n8n/config';
+import { entities } from '@n8n/db';
+import { Service } from '@n8n/di';
+import type { DataSourceOptions, LoggerOptions } from '@n8n/typeorm';
+import type { MysqlConnectionOptions } from '@n8n/typeorm/driver/mysql/MysqlConnectionOptions';
+import type { PostgresConnectionOptions } from '@n8n/typeorm/driver/postgres/PostgresConnectionOptions';
+import type { SqliteConnectionOptions } from '@n8n/typeorm/driver/sqlite/SqliteConnectionOptions';
+import type { SqlitePooledConnectionOptions } from '@n8n/typeorm/driver/sqlite-pooled/SqlitePooledConnectionOptions';
+import { InstanceSettings } from 'n8n-core';
+import { UserError } from 'n8n-workflow';
+import path from 'path';
+import type { TlsOptions } from 'tls';
+
+import { InsightsByPeriod } from '@/modules/insights/database/entities/insights-by-period';
+import { InsightsMetadata } from '@/modules/insights/database/entities/insights-metadata';
+import { InsightsRaw } from '@/modules/insights/database/entities/insights-raw';
+
+import { mysqlMigrations } from './migrations/mysqldb';
+import { postgresMigrations } from './migrations/postgresdb';
+import { sqliteMigrations } from './migrations/sqlite';
+import { subscribers } from './subscribers';
+
+@Service()
+export class DBConnectionOptions {
+	constructor(
+		private readonly globalConfig: GlobalConfig,
+		private readonly instanceSettings: InstanceSettings,
+	) {}
+
+	getOverrides(dbType: 'postgresdb' | 'mysqldb') {
+		const dbConfig = this.globalConfig.database[dbType];
+		return {
+			database: dbConfig.database,
+			host: dbConfig.host,
+			port: dbConfig.port,
+			username: dbConfig.user,
+			password: dbConfig.password,
+		};
+	}
+
+	getOptions(): DataSourceOptions {
+		const { type: dbType } = this.globalConfig.database;
+		switch (dbType) {
+			case 'sqlite':
+				return this.getSqliteConnectionOptions();
+			case 'postgresdb':
+				return this.getPostgresConnectionOptions();
+			case 'mariadb':
+			case 'mysqldb':
+				return this.getMysqlConnectionOptions(dbType);
+			default:
+				throw new UserError('Database type currently not supported', { extra: { dbType } });
+		}
+	}
+
+	private getCommonOptions() {
+		const { tablePrefix: entityPrefix, logging: loggingConfig } = this.globalConfig.database;
+
+		let loggingOption: LoggerOptions = loggingConfig.enabled;
+		if (loggingOption) {
+			const optionsString = loggingConfig.options.replace(/\s+/g, '');
+			if (optionsString === 'all') {
+				loggingOption = optionsString;
+			} else {
+				loggingOption = optionsString.split(',') as LoggerOptions;
+			}
+		}
+
+		return {
+			entityPrefix,
+			entities: [...Object.values(entities), InsightsRaw, InsightsByPeriod, InsightsMetadata],
+			subscribers: Object.values(subscribers),
+			migrationsTableName: `${entityPrefix}migrations`,
+			migrationsRun: false,
+			synchronize: false,
+			maxQueryExecutionTime: loggingConfig.maxQueryExecutionTime,
+			logging: loggingOption,
+		};
+	}
+
+	private getSqliteConnectionOptions(): SqliteConnectionOptions | SqlitePooledConnectionOptions {
+		const { sqlite: sqliteConfig } = this.globalConfig.database;
+		const { n8nFolder } = this.instanceSettings;
+
+		const commonOptions = {
+			...this.getCommonOptions(),
+			database: path.resolve(n8nFolder, sqliteConfig.database),
+			migrations: sqliteMigrations,
+		};
+
+		if (sqliteConfig.poolSize > 0) {
+			return {
+				type: 'sqlite-pooled',
+				poolSize: sqliteConfig.poolSize,
+				enableWAL: true,
+				acquireTimeout: 60_000,
+				destroyTimeout: 5_000,
+				...commonOptions,
+			};
+		} else {
+			return {
+				type: 'sqlite',
+				enableWAL: sqliteConfig.enableWAL,
+				...commonOptions,
+			};
+		}
+	}
+
+	private getPostgresConnectionOptions(): PostgresConnectionOptions {
+		const { postgresdb: postgresConfig } = this.globalConfig.database;
+		const {
+			ssl: { ca: sslCa, cert: sslCert, key: sslKey, rejectUnauthorized: sslRejectUnauthorized },
+		} = postgresConfig;
+
+		let ssl: TlsOptions | boolean = postgresConfig.ssl.enabled;
+		if (sslCa !== '' || sslCert !== '' || sslKey !== '' || !sslRejectUnauthorized) {
+			ssl = {
+				ca: sslCa || undefined,
+				cert: sslCert || undefined,
+				key: sslKey || undefined,
+				rejectUnauthorized: sslRejectUnauthorized,
+			};
+		}
+
+		return {
+			type: 'postgres',
+			...this.getCommonOptions(),
+			...this.getOverrides('postgresdb'),
+			schema: postgresConfig.schema,
+			poolSize: postgresConfig.poolSize,
+			migrations: postgresMigrations,
+			connectTimeoutMS: postgresConfig.connectionTimeoutMs,
+			ssl,
+		};
+	}
+
+	private getMysqlConnectionOptions(dbType: 'mariadb' | 'mysqldb'): MysqlConnectionOptions {
+		return {
+			type: dbType === 'mysqldb' ? 'mysql' : 'mariadb',
+			...this.getCommonOptions(),
+			...this.getOverrides('mysqldb'),
+			migrations: mysqlMigrations,
+			timezone: 'Z', // set UTC as default
+		};
+	}
+}
