+++ packages/cli/test/integration/shared/test-db.ts
@@ -5,8 +5,8 @@ import type { DataSourceOptions } from '@n8n/typeorm';
 import { DataSource as Connection } from '@n8n/typeorm';
 import { randomString } from 'n8n-workflow';
 
-import { getOptionOverrides } from '@/databases/config';
 import { DBConnection } from '@/databases/db-connection';
+import { DBConnectionOptions } from '@/databases/db-connection-options';
 
 export const testDbPrefix = 'n8n_test_';
 
@@ -74,7 +74,7 @@ export const getBootstrapDBOptions = (dbType: 'postgresdb' | 'mysqldb'): DataSou
 	const type = dbType === 'postgresdb' ? 'postgres' : 'mysql';
 	return {
 		type,
-		...getOptionOverrides(dbType),
+		...Container.get(DBConnectionOptions).getOverrides(dbType),
 		database: type,
 		entityPrefix: globalConfig.database.tablePrefix,
 		schema: dbType === 'postgresdb' ? globalConfig.database.postgresdb.schema : undefined,
